% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_asBruteMatrix.R
\name{as.brute.matrix}
\alias{as.brute.matrix}
\title{as.brute.matrix}
\usage{
as.brute.matrix(InteractionSer_object, first = "all", second = "all",
  fill = NULL, alternativeFill = FALSE, NAmin = FALSE,
  wkFolder = "")
}
\arguments{
\item{alternativeFill}{}
}
\description{
as.brute.matrix
}
\keyword{internal}
