% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_bamMatching.R
\name{bamMatching}
\alias{bamMatching}
\title{bamMatching}
\usage{
bamMatching(bam.datapath, param, h5name = "h5file", h5PathOut = "")
}
\arguments{
\item{h5PathOut}{}
}
\description{
bamMatching
}
\keyword{internal}
