% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_bamToDataframe.R
\name{bamToDataFrame}
\alias{bamToDataFrame}
\title{bamToDataframe}
\usage{
bamToDataFrame(bamPath, chromosome, start, end)
}
\arguments{
\item{end}{}
}
\description{
bamToDataframe
}
\keyword{internal}
