% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCK_vMod3_ICEnormalization2.R
\name{bck_diffHic_Normalization_Server}
\alias{bck_diffHic_Normalization_Server}
\title{diffHic_Normalization_Server}
\usage{
bck_diffHic_Normalization_Server(input, output, session, stringsAsFactors,
  wdPath)
}
\arguments{
\item{wdPath}{}
}
\description{
diffHic_Normalization_Server
}
\keyword{internal}
