% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCK_vMod3_ICEnormalization2.R
\name{bck_diffHic_Normalization_UI}
\alias{bck_diffHic_Normalization_UI}
\title{diffHic_Normalization_UI}
\usage{
bck_diffHic_Normalization_UI(id, label = "Visualization")
}
\arguments{
\item{label}{}
}
\description{
diffHic_Normalization_UI
}
\keyword{internal}
