% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_callModuleDescription.R
\name{callModuleDescription}
\alias{callModuleDescription}
\title{callmoduleDescription}
\usage{
callModuleDescription(mainNavItem)
}
\arguments{
\item{mainNavItem}{}
}
\description{
callmoduleDescription
}
\keyword{internal}
