% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod3_wavSisNormalization.R
\name{chromoR_Normalization_UI}
\alias{chromoR_Normalization_UI}
\title{chromoR_Normalization_UI}
\usage{
chromoR_Normalization_UI(id, label = "Normalization")
}
\arguments{
\item{label}{}
}
\description{
chromoR_Normalization_UI
}
\keyword{internal}
