% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_detectChromos.R
\name{detectMatrixCoordinate}
\alias{detectMatrixCoordinate}
\title{Title}
\usage{
detectMatrixCoordinate(path, onlyChr = FALSE)
}
\arguments{
\item{onlyChr}{}
}
\description{
Title
}
\keyword{internal}
