% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod2_Binning.R
\name{DiffHiC_BinningV2_Server}
\alias{DiffHiC_BinningV2_Server}
\title{DiffHiC_BinningV2_Server}
\usage{
DiffHiC_BinningV2_Server(input, output, session, stringsAsFactors, wdPath)
}
\arguments{
\item{wdPath}{}
}
\description{
DiffHiC_BinningV2_Server
}
\keyword{internal}
