% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod2_Binning.R
\name{DiffHiC_BinningV2_UI}
\alias{DiffHiC_BinningV2_UI}
\title{DiffHiC_BinningV2_UI}
\usage{
DiffHiC_BinningV2_UI(id, label = "binning")
}
\arguments{
\item{label}{}
}
\description{
DiffHiC_BinningV2_UI
}
\keyword{internal}
