% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod3_ICEnormalization.R,
%   R/vMod3_ICEnormalization.R
\name{diffHic_Normalization_Server}
\alias{diffHic_Normalization_Server}
\title{diffHic_Normalization_Server}
\usage{
diffHic_Normalization_Server(input, output, session, stringsAsFactors,
  wdPath)

diffHic_Normalization_Server(input, output, session, stringsAsFactors,
  wdPath)
}
\arguments{
\item{wdPath}{}

\item{wdPath}{}
}
\description{
diffHic_Normalization_Server

diffHic_Normalization_Server
}
\keyword{internal}
