% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod3_ICEnormalization.R,
%   R/vMod3_ICEnormalization.R
\name{diffHic_Normalization_UI}
\alias{diffHic_Normalization_UI}
\title{diffHic_Normalization_UI}
\usage{
diffHic_Normalization_UI(id, label = "Visualization")

diffHic_Normalization_UI(id, label = "Visualization")
}
\arguments{
\item{label}{}

\item{label}{}
}
\description{
diffHic_Normalization_UI

diffHic_Normalization_UI
}
\keyword{internal}
