% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod4_EpigeneticFeatures.R
\name{EpigeneticFeatures_Server}
\alias{EpigeneticFeatures_Server}
\title{pca_postProcessing_Server}
\usage{
EpigeneticFeatures_Server(input, output, session, stringsAsFactors, wdPath)
}
\arguments{
\item{wdPath}{}
}
\description{
pca_postProcessing_Server
}
\keyword{internal}
