% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_extractContact.R
\name{extractContact}
\alias{extractContact}
\title{extractContact}
\usage{
extractContact(interactionSetObj, intChrFirst, intChrSecond)
}
\arguments{
\item{intChrSecond}{}
}
\description{
extractContact
}
\keyword{internal}
