% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod1_filtering.R
\name{FilmodServer}
\alias{FilmodServer}
\title{FilmodServer}
\usage{
FilmodServer(input, output, session, stringsAsFactors, wdPath)
}
\arguments{
\item{wdPath}{}
}
\description{
FilmodServer
}
\keyword{internal}
