% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_findRegionsInBint.R
\name{findRegionInBint}
\alias{findRegionInBint}
\title{findRegionInBint}
\usage{
findRegionInBint(bint.bed_datapath, chr, start, end, id = TRUE,
  twoReg = FALSE, secChr, secStart, secEnd, imported = FALSE,
  only.chr = FALSE)
}
\arguments{
\item{only.chr}{}
}
\description{
findRegionInBint
}
\keyword{internal}
