% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_folderFrame.R
\name{folderFrame}
\alias{folderFrame}
\title{folderFrame}
\usage{
folderFrame(path)
}
\arguments{
\item{path}{path from which to obtain the data frame}
}
\description{
obtain data frame with list of folder in specific path
}
\keyword{internal}
