% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_folderList.R
\name{folderList}
\alias{folderList}
\title{folderList}
\usage{
folderList(path, type = "dirs", ecception = FALSE, ecceptionName)
}
\arguments{
\item{ecceptionName}{}
}
\description{
folderList
}
\keyword{internal}
