% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_hcrBck.R
\name{hcrBck}
\alias{hcrBck}
\title{hcrBck}
\usage{

  hcrBck(pathFrom = "/media/lucio/FastData/3.SuperGrive/Grive/1.Bioinformatic/2.Rprj/2.devel/",
  pathTo = "/media/lucio/data/10.bck/HCRbck/", paName = "HiCeekR")
}
\arguments{
\item{paName}{}
}
\description{
hcrBck
}
\keyword{internal}
