% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_HCRobserveEv.R
\name{HCRobserveEvent}
\alias{HCRobserveEvent}
\title{HCRobserveEvent}
\usage{
HCRobserveEvent(eventExpr, handlerExpr, event.env = parent.frame(),
  event.quoted = FALSE, handler.env = parent.frame(),
  handler.quoted = FALSE, label = NULL, suspended = FALSE,
  priority = 0, domain = getDefaultReactiveDomain(),
  autoDestroy = TRUE, ignoreNULL = TRUE, ignoreInit = FALSE,
  once = FALSE)
}
\arguments{
\item{once}{}
}
\description{
HCRobserveEvent
}
\keyword{internal}
