% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_runApp.R
\name{HCRprjManager}
\alias{HCRprjManager}
\title{HCRprjManager}
\usage{
HCRprjManager(configFilePath = "./HCR.config")
}
\arguments{
\item{configFilePath}{}
}
\description{
HCRprjManager
}
\keyword{internal}
