% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_HCRwrite&read.R
\name{HCRread}
\alias{HCRread}
\title{HCRread}
\usage{
HCRread(file, path, header = TRUE, sep = "\\t", quote = "\\"'",
  dec = ".", numerals = c("allow.loss", "warn.loss", "no.loss"), row.names,
  col.names, as.is = !stringsAsFactors, na.strings = "NA",
  colClasses = NA, nrows = -1, skip = 0, check.names = TRUE,
  fill = !blank.lines.skip, strip.white = FALSE, blank.lines.skip = TRUE,
  comment.char = "#", allowEscapes = FALSE, flush = FALSE,
  stringsAsFactors = default.stringsAsFactors(), fileEncoding = "",
  encoding = "unknown", text, skipNul = FALSE, asMatrix = TRUE)
}
\arguments{
\item{asMatrix}{}
}
\description{
HCRread
}
\keyword{internal}
