% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_HCRwrite&read.R
\name{HCRwrite}
\alias{HCRwrite}
\title{HCRwrite}
\usage{
HCRwrite(tableToWrite, file, path = "", append = FALSE, quote = FALSE,
  sep = "\\t", eol = "\\n", na = "NA", dec = ".", row.names = FALSE,
  col.names = TRUE, qmethod = c("escape", "double"), fileEncoding = "",
  standard.row.names = FALSE, extension = TRUE)
}
\arguments{
\item{extension}{}
}
\description{
HCRwrite
}
\keyword{internal}
