% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod4_HMMdownstream.R
\name{HMM_postProcessing_Server}
\alias{HMM_postProcessing_Server}
\title{HMM_postProcessing_Server}
\usage{
HMM_postProcessing_Server(input, output, session, stringsAsFactors, wdPath)
}
\arguments{
\item{wdPath}{}
}
\description{
HMM_postProcessing_Server
}
\keyword{internal}
