% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod4_HMMdownstream.R
\name{HMM_postProcessing_UI}
\alias{HMM_postProcessing_UI}
\title{HMM_postProcessing_UI}
\usage{
HMM_postProcessing_UI(id, label = "TADs")
}
\arguments{
\item{label}{}
}
\description{
HMM_postProcessing_UI
}
\keyword{internal}
