% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_importChrFromFragList.R
\name{importChrFromFragList}
\alias{importChrFromFragList}
\title{importChrFromFragList}
\usage{
importChrFromFragList(regPath)
}
\arguments{
\item{regPath}{}
}
\description{
importChrFromFragList
}
\keyword{internal}
