% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_makeBranch.R
\name{makeBranch}
\alias{makeBranch}
\title{makeBranch}
\usage{
makeBranch(branchName, pipelineStep, configFile = "./HCR.config",
  infoTsv)
}
\arguments{
\item{infoTsv}{}
}
\description{
makeBranch
}
\keyword{internal}
