% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_prjManager.R
\name{makeConfig}
\alias{makeConfig}
\title{makeConfig}
\usage{
makeConfig(path = paste0(getwd(), "/"))
}
\arguments{
\item{path}{character vector that contain folder path}
}
\value{
HCRwd.info file that contain the path of HiCeekR woorking directory
}
\description{
make HiCeekR config file
}
\examples{

}
\keyword{internal}
