% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_prjManager.R
\name{makeHCRan}
\alias{makeHCRan}
\title{makeHCRan}
\usage{
makeHCRan(name, prjPath, infoTable)
}
\arguments{
\item{name}{character varible that contain the name of analysis}

\item{prjPath}{character variable that contain the path of projects that}

\item{infoTable}{table that contai analysis information}
}
\description{
makeHCRan
}
\keyword{internal}
