% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_prjManager.R
\name{makeHCRprj}
\alias{makeHCRprj}
\title{make HiCeekR project folder}
\usage{
makeHCRprj(name, path)
}
\arguments{
\item{name}{character vector that defines name of project}

\item{path}{character vector that contain folder path}
}
\value{
None
}
\description{
make Project folder in which HiCeekR works
}
\examples{

}
\keyword{internal}
