% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_prjManager.R
\name{makeHCRwd}
\alias{makeHCRwd}
\title{makeHiCeekRwd}
\usage{
makeHCRwd(path = getwd())
}
\arguments{
\item{path}{character vector that contain folder path}
}
\value{
None
}
\description{
make folder in which HiCeekR works
}
\examples{

}
