% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_matrixExport.R
\name{matrixExport}
\alias{matrixExport}
\title{matrixExport}
\usage{
matrixExport(redata, type = "all", chrA, chrB, chgFill = FALSE,
  alternativeFill, NAas0 = TRUE)
}
\arguments{
\item{NAas0}{}
}
\description{
matrixExport
}
\keyword{internal}
