% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_matrixWithNames.R
\name{matrixWithNames}
\alias{matrixWithNames}
\title{matrixWithNames}
\usage{
matrixWithNames(matrixWithIdPath, bint.bed_Path, speed = FALSE, chromoA,
  chromoB, bedPath = TRUE)
}
\arguments{
\item{bedPath}{}
}
\description{
matrixWithNames
}
\keyword{internal}
