% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_xPanels_moduleStartPanel.R
\name{moduleStartPanel}
\alias{moduleStartPanel}
\title{moduleStartPanel}
\usage{
moduleStartPanel(mainNavItem, wdPath, moduleDescription)
}
\arguments{
\item{moduleDescription}{}
}
\description{
moduleStartPanel
}
\keyword{internal}
