% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod0_moduleTemplate.R
\name{moduleTemplate_Server}
\alias{moduleTemplate_Server}
\title{moduleTemplate_Server}
\usage{
moduleTemplate_Server(input, output, session)
}
\arguments{
\item{session}{}
}
\description{
moduleTemplate_Server
}
\keyword{internal}
