% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod0_moduleTemplate.R
\name{moduleTemplate_UI}
\alias{moduleTemplate_UI}
\title{moduleTemplate_UI}
\usage{
moduleTemplate_UI(id, label = "moduleTemplate")
}
\arguments{
\item{label}{}
}
\description{
moduleTemplate_UI
}
\keyword{internal}
