% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_runApp.R
\name{moduleTest}
\alias{moduleTest}
\title{moduleTest}
\usage{
moduleTest(moduleName)
}
\arguments{
\item{moduleName}{character variable that contain name of module}
}
\value{
shinyApp in web browser
}
\description{
Test the module without HiCeekR
}
\keyword{internal}
