% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod4_PCAdownstream.R
\name{pca_postProcessing_Server}
\alias{pca_postProcessing_Server}
\title{pca_postProcessing_Server}
\usage{
pca_postProcessing_Server(input, output, session, stringsAsFactors, wdPath)
}
\arguments{
\item{wdPath}{}
}
\description{
pca_postProcessing_Server
}
\keyword{internal}
