% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod4_EpigeneticFeatures.R
\name{pca_postProcessing_UI}
\alias{pca_postProcessing_UI}
\title{pca_postProcessing_UI}
\usage{
pca_postProcessing_UI(id, label = "epiFea")
}
\arguments{
\item{label}{}
}
\description{
pca_postProcessing_UI
}
\keyword{internal}
