% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod4_PCAdownstream.R
\name{pcaComp_postProcessing_Server}
\alias{pcaComp_postProcessing_Server}
\title{pcaComp_postProcessing_Server}
\usage{
pcaComp_postProcessing_Server(input, output, session, stringsAsFactors,
  wdPath)
}
\arguments{
\item{wdPath}{}
}
\description{
pcaComp_postProcessing_Server
}
\keyword{internal}
