% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod4_PCAdownstream.R
\name{pcaComp_postProcessing_UI}
\alias{pcaComp_postProcessing_UI}
\title{pcaComp_postProcessing_UI}
\usage{
pcaComp_postProcessing_UI(id, label = "compartments")
}
\arguments{
\item{label}{}
}
\description{
pcaComp_postProcessing_UI
}
\keyword{internal}
