% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFuns_PCA.R
\name{pcaMatrixModel}
\alias{pcaMatrixModel}
\title{pcaMatrixModel}
\usage{
pcaMatrixModel(binTablePath, columnNumber = 1)
}
\arguments{
\item{columnNumber}{}
}
\description{
pcaMatrixModel
}
\keyword{internal}
