% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_pointin.R
\name{pointin}
\alias{pointin}
\title{pointin}
\usage{
pointin(analysisFolder, specificFolderName, sys = FALSE)
}
\arguments{
\item{analysisFolder}{Character vector with path of specific analysis folder}

\item{specificFolderName}{Character vector with name of folder of interest}

\item{sys}{bulean fro choicing if save in Results folder or in SysOut folder}
}
\description{
pointin
}
\keyword{internal}
