% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_predictTADs.R
\name{predictTADs}
\alias{predictTADs}
\title{predictTADs}
\usage{
predictTADs(hmmPath, tollerance = 3, maxspan = 3, spanMark = FALSE,
  saveBed = NULL)
}
\arguments{
\item{saveBed}{}
}
\description{
predictTADs
}
\keyword{internal}
