% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_xPanels_prjManager.R
\name{prjManPanel_analysisSettings}
\alias{prjManPanel_analysisSettings}
\title{prjManPanel_analysisSettings}
\usage{
prjManPanel_analysisSettings()
}
\description{
prjManPanel_analysisSettings
}
\keyword{internal}
