% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_xPanels_prjManager.R
\name{prjManPanel_anlysisInputType}
\alias{prjManPanel_anlysisInputType}
\title{prjManPanel_analysisInputType}
\usage{
prjManPanel_anlysisInputType()
}
\description{
prjManPanel_analysisInputType
}
\keyword{internal}
