% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_xPanels_prjManager.R
\name{prjManPanel_inputBAM}
\alias{prjManPanel_inputBAM}
\title{prjManPanel_inputBAM}
\usage{
prjManPanel_inputBAM(prjFolder, input)
}
\arguments{
\item{input}{}
}
\description{
prjManPanel_inputBAM
}
\keyword{internal}
