% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_xPanels_prjManager.R
\name{prjManPanel_loadProject}
\alias{prjManPanel_loadProject}
\title{prjManPanel_loadProject}
\usage{
prjManPanel_loadProject(workingDir)
}
\arguments{
\item{workingDir}{}
}
\description{
prjManPanel_loadProject
}
\keyword{internal}
