% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod1_projectSettings.R
\name{prjSettings_Server}
\alias{prjSettings_Server}
\title{prjSettings_Server}
\usage{
prjSettings_Server(input, output, session, prjManExport)
}
\arguments{
\item{prjManExport}{}
}
\description{
prjSettings_Server
}
\keyword{internal}
