% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod1_projectSettings.R
\name{prjSettings_UI}
\alias{prjSettings_UI}
\title{prjSettings_UI}
\usage{
prjSettings_UI(id, label = "prjSettings")
}
\arguments{
\item{label}{}
}
\description{
prjSettings_UI
}
\keyword{internal}
