% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_refGenomeEdit.R
\name{refGenomeEdit}
\alias{refGenomeEdit}
\title{refGenomeEdit}
\usage{
refGenomeEdit(refGenome, enzyme, overhang, name = "refGenFrags",
  cutGenomeOutDir = "")
}
\arguments{
\item{cutGenomeOutDir}{}
}
\description{
refGenomeEdit
}
\keyword{internal}
