% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_regionsInMatrix.R
\name{regionsInMatrix}
\alias{regionsInMatrix}
\title{regionsInMatrix}
\usage{
regionsInMatrix(matrixPath, separate = FALSE, on = "row",
  unique = TRUE, as.Path = TRUE)
}
\arguments{
\item{as.Path}{}
}
\description{
regionsInMatrix
}
\keyword{internal}
