% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_restartFrom.R
\name{restartFrom}
\alias{restartFrom}
\title{restartFrom}
\usage{
restartFrom(infoTablePath, pipelineStep, configFile = "./HCR.config")
}
\arguments{
\item{configFile}{}
}
\description{
restartFrom
}
\keyword{internal}
