% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfun_heatmaply_searchTick.R
\name{searchTicks}
\alias{searchTicks}
\title{searchTicks}
\usage{
searchTicks(matrix, axis = "row", numberOfTicks = 14, as.path = FALSE)
}
\arguments{
\item{as.path}{}
}
\description{
searchTicks
}
\keyword{internal}
