% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_selectFile.R
\name{selectFile}
\alias{selectFile}
\title{selectFile}
\usage{
selectFile(id, path, label = "Select File", subset = FALSE,
  pattern = "NA", width = "100\%")
}
\arguments{
\item{width}{}
}
\description{
selectFile
}
\keyword{internal}
