% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_prjManager.R
\name{setHCRwd}
\alias{setHCRwd}
\title{setHCRwd}
\usage{
setHCRwd(path)
}
\arguments{
\item{path}{character vector that contain folder path}
}
\value{
HCRwd.info file that contain the path of HiCeekR woorking directory
}
\description{
set folder in which HiCeekR works
}
\examples{

}
