% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod5_heatmapVisualization.R
\name{TADsV2_Visualization_Server}
\alias{TADsV2_Visualization_Server}
\title{TADsV2_Visualization_Server}
\usage{
TADsV2_Visualization_Server(input, output, session, stringsAsFactors,
  wdPath)
}
\arguments{
\item{wdPath}{}
}
\description{
TADsV2_Visualization_Server
}
\keyword{internal}
