% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod5_heatmapVisualization.R
\name{TADsV2_Visualization_UI}
\alias{TADsV2_Visualization_UI}
\title{TADsV2_Visualization_UI}
\usage{
TADsV2_Visualization_UI(id, label = "Visualization")
}
\arguments{
\item{label}{}
}
\description{
TADsV2_Visualization_UI
}
\keyword{internal}
