% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_testInput.R
\name{testInfoTable}
\alias{testInfoTable}
\title{testInfoTable}
\usage{
testInfoTable(infoTable, stop = TRUE)
}
\arguments{
\item{infoTable}{the table that contain analysis information}

\item{stop}{stop process if test fail}
}
\value{
boolean TRUE if input is correct
}
\description{
test if infoTable is correct
}
\keyword{internal}
