% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_heatmaply_tickOptions.R
\name{tickOptions}
\alias{tickOptions}
\title{tickOptions}
\usage{
tickOptions(tickNamesObj, chr = TRUE, start = TRUE, end = FALSE)
}
\arguments{
\item{end}{}
}
\description{
tickOptions
}
\keyword{internal}
