% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFun_heatmaply_tickNames.R
\name{ticksNames}
\alias{ticksNames}
\title{ticksNames}
\usage{
ticksNames(ticks, matrix, axis = "row", as.path = FALSE, revert = FALSE,
  simMirror = FALSE)
}
\arguments{
\item{simMirror}{}
}
\description{
ticksNames
}
\keyword{internal}
