% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMod3_ICEnormalization.R
\name{diffHic_Normalization_Server}
\alias{diffHic_Normalization_Server}
\title{diffHic_Normalization_Server}
\usage{
diffHic_Normalization_Server(input, output, session, stringsAsFactors,
  wdPath)
}
\arguments{
\item{wdPath}{}
}
\description{
diffHic_Normalization_Server
}
\keyword{internal}
